KAR-400 Fixed up the unit tests
In the happy_path test we had forgotten to add the assert keyword, and hence were getting
false positive. With that change I also added time.sleep to allow the component manager's
background thread to have enough time to update the state before we check it.

In the non_existing_component test I have mock patched the DeviceProxy to get rid of the
error about the TANGO_HOST env not set. I removed the check for CommunicationStatus == DISABLED because
when the task is submitted, it is queued and we set the communication state to NOT_ESTABLISHED immediately. So

Co-authored-by: Samuel Twum <samueltwum1@gmail.com> , Phelokazi Dube <pdube@ska.ac.za>
Signed-off-by: Katleho Madisa <katleho.madisa47@gmail.com>