KAR-455 Added the DS achievedPointing attr to be monitored
This adds the DS achievedPointing attribute to the list of monitored attributes by the component manager.The value from this
monitored attribute is used to update the achievedPointing attribute on the DishManager. This is a workaround to not
using Tango  forwarded attributes.

Attrbutes of array types usually return the values as numpy arrays, which the _update_component_state method in the ska-tango-base doesn't
handle when checking for state changes. It raises this error, ValueError: The truth value of an array with more than one element is ambiguous. Use a.any() or a.all().
So to work around this we convert the numpy array to a list.

Signed-off-by: Katleho Madisa <katleho.madisa47@gmail.com>