KAR-687 Rework branching condition in while loop
Some of the sub commands do not require dish manager to wait for any changes. The while
loop terminates when we skip progress updates for such commands. The previous MR #101
introduced a bug which makes all commands to terminate causing updates needed for some
commands not to come through. The boolean operation has been reworked to fix this